fixed bug in assert message, tried tightloop congruence
the assert message did not receive the variadic argument list, so it just printed random bytes. Tightloop is an attempt to only do gate extraction once, and from then on only continue refininig and simplifying the existing gates. created some problems though where the program crashes. a bit more robust is to hoist the loop outwards and call kissat_congruence on each iteration, eventhough then it needs to re-discover all the gates